+++ src/grib_coder/sections/grib_section_7.cpp
@@ -22,7 +22,7 @@ bool GribSection7::parseFile(std::FILE* file, bool header_only) {
     } else {
         std::vector<unsigned char> buffer(section_length_);
         const auto read_count = std::fread(&buffer[5], 1, buffer_length, file);
-        if (read_count != buffer_length) {
+        if (static_cast<long>(read_count) != buffer_length) {
             return false;
         }
